requirements.txt
certifi==2020.6.20
chardet==3.0.4
dnspython==1.16.0
eventlet==0.26.1
greenlet==0.4.16
idna==2.10
monotonic==1.5
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.24.0
six==1.15.0
urllib3==1.25.10
websocket-client==0.57.0
